backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
celery>=5.3
django-celery-results>=2.5
redis>=5.0
gunicorn>=21.2
psycopg2-binary>=2.9
Pillow>=11.0
